+++ src/v2rayF.Android/Application.cs
@@ -22,6 +22,7 @@ public class V2rayApplication : AvaloniaAndroidApplication<v2rayF.App>
         AppServices.CoreEnvironment = new AndroidCoreEnvironment();
         AppServices.Platform = new AndroidPlatformIntegration();
         AppServices.CoreProcessHost = new AndroidJavaCoreProcessHost();
+        AppServices.Updater = new AndroidAppUpdater();
         AndroidEnvironment.UnhandledExceptionRaiser += (_, e) =>
             global::Android.Util.Log.Error("v2rayF", e.Exception?.ToString() ?? "Unhandled exception");
         base.OnCreate();